Split ReinforcedWallVisualizer into Component/System pair
