+++ Content.Client/Wall/ReinforcedWallVisualizerComponent.cs
@@ -1,0 +1,6 @@
+namespace Content.Client.Wall;
+
+[RegisterComponent]
+[Access(typeof(ReinforcedWallVisualizerSystem))]
+public sealed class ReinforcedWallVisualizerComponent : Component
+{}
